moved set* methods to base
