+++ src/gtk/toga_gtk/widgets/label.py
@@ -25,24 +25,6 @@ class Label(Widget):
         self.native.set_xalign(values[0])
         self.native.set_yalign(values[1])
 
-    def set_color(self, color):
-        if color:
-            style_context = self.native.get_style_context()
-            css = get_color_css(color)
-            apply_gtk_style(style_context, css, "toga-color")
-
-    def set_background_color(self, color):
-        if color:
-            style_context = self.native.get_style_context()
-            css = get_bg_color_css(color)
-            apply_gtk_style(style_context, css, "toga-bg-color")
-
-    def set_font(self, value):
-        if value:
-            style_context = self.native.get_style_context()
-            css = get_font_css(value)
-            apply_gtk_style(style_context, css, "toga-font")
-
     def set_text(self, value):
         # FIXME after setting the label the label jumps to the top left
         # corner and only jumps back at its place after resizing the window.
